problem_fixed

+++ monies.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    unsigned int dollars = 4294967298;
+    unsigned long dollars = 4294967298;
     cout << "I have " << dollars << " dollars in my wallet." << endl;
     return 0;
 }
